+++ kubejs/client_scripts/mod_specific/hearth_and_harvest.js
@@ -1,0 +1,7 @@
+(() => {
+  ClientEvents.generateAssets("after_mods", (event) => {
+    event.json("emi:index/stacks/hearth_and_harvest_potion_effects", {
+      filters: ["hearthandharvest:pungent", "hearthandharvest:tempting"],
+    });
+  });
+})();